Using i64 instead of i32

+++ src/28/hello_rayon/benches/my_benchmark.rs
@@ -7,8 +7,8 @@ use hello_rayon::join_test;
 use hello_rayon::sum_array_iterative;
 use hello_rayon::sum_array_parallel;
 
-fn create() -> Vec<i32> {
-    let mut input: Vec<i32> = vec![];
+fn create() -> Vec<i64> {
+    let mut input: Vec<i64> = vec![];
     for n in 0..10000000 {
         input.push(n);
     }
@@ -16,21 +16,21 @@ fn create() -> Vec<i32> {
 }
 
 fn criterion_benchmark(c: &mut Criterion) {
-    let mut input: Vec<i32> = create();
+    let mut input: Vec<i64> = create();
     c.bench_function("sum_array_iterative", move |b| {
         b.iter(|| sum_array_iterative(&mut input))
     });
 }
 
 fn criterion_benchmark2(c: &mut Criterion) {
-    let mut input: Vec<i32> = create();
+    let mut input: Vec<i64> = create();
     c.bench_function("sum_array_parallel", move |b| {
         b.iter(|| sum_array_parallel(&mut input))
     });
 }
 
 fn criterion_benchmark3(c: &mut Criterion) {
-    let mut input: Vec<i32> = create();
+    let mut input: Vec<i64> = create();
     let (first_part, second_part) = input.split_at(input.len() / 2);
     let mut first_part = first_part.to_vec();
     let mut second_part = second_part.to_vec();
